requirements.txt
apispec == 6.0.0
apispec-webframeworks == 0.5.2
bcrypt == 4.0.1
celery[redis] == 5.2.7
flask-apispec == 0.11.4
Flask == 2.2.2
Flask-Cors == 3.0.10
Flask-RESTful == 0.3.9
Flask-SQLAlchemy == 3.0.2
gunicorn == 20.1.0
hiredis == 2.0.0
marshmallow == 3.18.0
pip == 22.3
psycopg2-binary == 2.9.4
pycodestyle == 2.9.1
PyYAML == 5.4.1
redis == 4.3.4
setuptools == 65.5.0
SQLAlchemy == 1.4.42
ujson == 5.5.0
webargs == 8.2.0
